Catch NaN values in histogram incrementing
In newhybrids with lots of SNPs I ran into a situation where histogram
incrementing was causing seg faults because the values were NaNs.

I added a little catch in there.  It's a little dodgy, since there
should not be NaN's there in the first place and I am concerned about
that, however, I am hoping that they are only there for the first few
iterations, and then it will get better.

This closes issue https://github.com/eriqande/newhybrids/issues/1